fix: Support empty entry JS in build mode.
