+++ examples/rust-vite/rust-host/package.json
@@ -4,7 +4,7 @@
   "scripts": {
     "dev": "rsbuild dev",
     "build": "rsbuild build",
-    "preview": "npm run build && rsbuild"
+    "preview": "npm run build && rsbuild preview"
   },
   "dependencies": {
     "@module-federation/enhanced": "0.2.5",
